+++ src/Template.js
@@ -42,17 +42,17 @@ class Template {
     render(params = {}) {
         let content = this.#content;
         const tagPattern = new RegExp(
-            '(?<indent>[ \\t]*)(?<placeholder>(?<prefix>[\\' +
-                Object.values(Prefix).join('\\') +
-                '])\\' +
+            '(?<indent>[ \\t]*)(?<placeholder>\\' +
+                Prefix.PARAMETER +
+                '\\' +
                 Bracket.OPEN +
                 '(?<path>[^\\' +
                 Bracket.CLOSE +
                 '\\|]*)(\\' +
                 Glue.FILTER +
-                '(?<filter>[^\\' +
+                '(?<filters>[^\\' +
                 Bracket.CLOSE +
-                '\\|]+))?\\' +
+                ']+))?\\' +
                 Bracket.CLOSE +
                 ')',
             'm',
@@ -60,13 +60,13 @@ class Template {
         let tag;
         while ((tag = tagPattern.exec(content))) {
             let param = this.#paramFromPath(tag.groups.path, params);
-            if (tag.groups.filter) {
-                param = this.#applyFilter(tag.groups.filter, param);
+            if (tag.groups.filters) {
+                const filters = tag.groups.filters.split(Glue.FILTER);
+                for (const filter of filters) {
+                    param = this.#applyFilter(filter, param);
+                }
             }
             if (typeof param === 'string') {
-                if (tag.groups.prefix === Prefix.ESCAPED) {
-                    param = this.#escape(param);
-                }
                 if (tag.groups.indent && param.includes('\n')) {
                     param = param.replace(new RegExp('\\n', 'g'), '\n' + tag.groups.indent);
                 }
@@ -93,6 +93,9 @@ class Template {
             return param.at(0).toUpperCase() + param.slice(1);
         case StringFilter.TRIM:
             return param.trim();
+        case StringFilter.ESCAPE:
+        case StringFilter.ESCAPE_SHORT:
+            return this.#escape(param);
         case NumberFilter.ABS:
             return Math.abs(parseFloat(param));
         case NumberFilter.ROUND:
@@ -145,10 +148,13 @@ class Template {
     }
 
     /**
-     * @param {String} text
-     * @return {String}
+     * @param {*} text
+     * @return {*}
      */
     #escape(text) {
+        if (typeof text !== 'string') {
+            return text;
+        }
         return text
             .replace(/([<>&]|[^#-~| |!])/g, (match) => '&#' + match.charCodeAt(0) + ';')
             .replace(
@@ -193,7 +199,7 @@ class Template {
             const blockId = 'block_' + counter;
             counter++;
             const blockPlaceholder =
-                Prefix.RAW + Bracket.OPEN + Operator.PARTIAL + blockId + Bracket.CLOSE;
+                Prefix.PARAMETER + Bracket.OPEN + Operator.PARTIAL + blockId + Bracket.CLOSE;
             this.#content = this.#content.replace(block[0], blockPlaceholder);
             blockPartials[blockId] = this.#blockCallback(block);
         }